requirements.txt
Flask>=3.0.0
requests>=2.31.0
torch>=2.2.0
torchvision>=0.17.0
transformers==4.40.0
pillow>=10.1.0
accelerate>=0.29.3
sentencepiece
protobuf>=3.20.3
sentence-transformers==2.5.1
numpy
chromadb>=0.4.22
einops

